# swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample Unicef server."
  version: "1.0.0"
  title: "Unicef"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: ""
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
  /api/view/files:
    get:
      tags:
      - "Medications"
      summary: "Medication file list"
      description: "This api will give all catalougue file list"
      operationId: "fetchMedications"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      responses:
        "200":
           description: "Invalid input"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "Files found!!"
               data:
                 type: "array"
                 items:
                   $ref: "#/definitions/Cataloguefile"
        "500":
           description: "Invalid input"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "Files not found!!"
               data:
                 type: "array"
                 items:
                   type: string
                   example: null
...
  /api/upload/medications:
    post:
      tags:
      - "Medications"
      summary: "uploads catalougue file"
      description: ""
      operationId: "uploadFile"
      consumes:
      - "multipart/form-data"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "userId"
        in: "formData"
        description: "ID of User"
        required: true
        type: "string"
        example: "6048708c946251d02c88f329"
      - name: "supplierCode"
        in: "formData"
        description: "supplierCode"
        example: "ALKEM"
        required: true
        type: "string"
      - name: "isoCountryCode"
        in: "formData"
        description: "isoCountryCode"
        required: true
        type: "string"
        example: "IND"
      - name: "file"
        in: "formData"
        description: "file to upload"
        required: true
        type: "file"
      responses:
         "200":
           description: "successful operation"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: true
               message:
                 type: "string"
                 example: "Data Inserted Successfully"
               invalidRowsCount:
                 type: "integer"
                 example: 25
               validRowsCount:
                 type: "integer"
                 example: 50
               totalRowsCount:
                 type: "integer"
                 example: 100
               duplicateEntryCount:
                 type: "integer"
                 example: 25
         "500":
           description: "Invalid input"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "Error in data insert process"
               invalidRowsCount:
                 type: "integer"
                 example: 0
               validRowsCount:
                 type: "integer"
                 example: 0
               totalRowsCount:
                 type: "integer"
                 example: 0
               duplicateEntryCount:
                 type: "integer"
                 example: 0
                 
         "400":
           description: "parameter is missing"
           schema:
             type: "object"
             properties:
               status:
                 type: "boolean"
                 example: false
               message:
                 type: "string"
                 example: "supplierCode parameter is missing"



# schemas
definitions:
  Cataloguefile:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "ObjectId"
      fileName:
        type: "string"
       
      userId:
        type: "string"
       
      supplierCode:
        type: "string"
      successedRecordsCount:
        type: "integer"
        format: "int32"
      duplicateRecordsCount:
        type: "integer"
        format: "int32"
      failedRecordsCount:
        type: "integer"
        format: "int32"
      totalRecordsCount:
        type: "integer"
        format: "int32"
      status:
        type: "boolean"
        default: false
      isoCountryCode:
         type: "string"
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
    xml:
      name: "Cataloguefile"
...